refact: move as actions do dispatch do contexto pra um novo arquivo
